backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
matplotlib
requests
python-multipart
anthropic
python-dotenv
yfinance
joblib
